requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
weasyprint==61.2
jinja2==3.1.3
python-dateutil==2.9.0
